pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "catalyst-radar"
version = "0.1.0"
description = "Deterministic-first market radar for public-equity opportunity review"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "numpy>=1.26",
  "pandas>=2.2",
  "psycopg[binary]>=3.2",
  "python-dotenv>=1.0",
  "sqlalchemy>=2.0",
  "streamlit>=1.35",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "ruff>=0.5",
]

[project.scripts]
catalyst-radar = "catalyst_radar.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
